+++ datadope_alerta/bgtasks/__init__.py
@@ -11,8 +11,8 @@ from alerta.utils.collections import merge  # To provide import for package modu
 # noinspection PyUnresolvedReferences
 from alertaclient.api import Client as AlertaClient  # To provide import for package modules
 
-from iometrics_alerta import initialize, is_initialized
-from iometrics_alerta.plugins import getLogger
+from datadope_alerta import initialize, is_initialized
+from datadope_alerta.plugins import getLogger
 
 
 if not is_initialized():
@@ -36,6 +36,8 @@ def revoke_task(task_id):
 from .periodic_tasks import check_automatic_closing # noqa - To provide import for package modules
 
 from .recovery_actions import launch_actions  # noqa - To provide import for package modules
+
+from .async_alert_task import async_receive  # noqa - To provide import for package modules
 
 # Tasks defined as classes must be instantiated and registered
 from .alert import event_task, recovery_task, repeat_task, action_task  # noqa - To provide import for package modules