+++ datadope_alerta/bgtasks/recovery_actions.py
@@ -11,13 +11,13 @@ from requests.exceptions import ConnectionError as RequestsConnectionError, Time
 from alerta.exceptions import AlertaException
 from alerta.models.enums import Status
 
-from iometrics_alerta import DateTime, RecoveryActionsFields, thread_local
-from iometrics_alerta import GlobalAttributes
-from iometrics_alerta.backend.flexiblededup.models.recovery_actions import RecoveryActionData, RecoveryActionsStatus
-from iometrics_alerta.plugins import Alerter, RetryableException
+from datadope_alerta import DateTime, RecoveryActionsFields, thread_local
+from datadope_alerta import GlobalAttributes
+from datadope_alerta.backend.flexiblededup.models.recovery_actions import RecoveryActionData, RecoveryActionsStatus
+from datadope_alerta.plugins import Alerter, RetryableException
 from . import app, celery, getLogger, Alert
 from . import revoke_task  # noqa - Provide import to other classes
-from ..recovery_actions.providers import RecoveryActionsProvider, RecoveryActionsResponseStatus, RecoveryActionsResponse
+from datadope_alerta.plugins.recovery_actions.providers import RecoveryActionsProvider, RecoveryActionsResponseStatus, RecoveryActionsResponse
 
 logger = getLogger(__name__)